Switched profile and home tabs
