+++ app/src/main/res/layout/fragment_student_home.xml
@@ -1,6 +1,76 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
     android:layout_height="match_parent">
+
+    <TextView
+        android:id="@+id/upcoming_meetings"
+        android:layout_width="192dp"
+        android:layout_height="64dp"
+        android:layout_marginTop="4dp"
+        android:text="Upcoming Meetings"
+        android:textSize="24sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.543"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <LinearLayout
+        android:id="@+id/tutors"
+        android:layout_width="match_parent"
+        android:layout_height="150dp"
+        android:layout_marginTop="16dp"
+        android:orientation="vertical"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/upcoming_meetings">
+
+    </LinearLayout>
+
+    <Button
+        android:id="@+id/button_view_more"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="View more"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        tools:layout_editor_absoluteY="280dp"
+        app:layout_constraintTop_toBottomOf="@id/tutors"/>
+
+    <TextView
+        android:id="@+id/your_subjects2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="60dp"
+        android:text="Your Subjects"
+        android:textSize="24sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/button_view_more" />
+
+    <LinearLayout
+        android:id="@+id/subjects2"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.6"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/your_subjects2"
+        app:layout_constraintVertical_bias="0.217">
+
+        <!--        <TextView-->
+        <!--            android:layout_width="88dp"-->
+        <!--            android:layout_height="43dp"-->
+        <!--            android:layout_marginStart="20dp"-->
+        <!--            android:layout_marginBottom="8dp"-->
+        <!--            android:text="Math" />-->
+
+
+    </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
