+++ app/src/main/res/layout/fragment_tutor_home.xml
@@ -33,13 +33,13 @@
     <LinearLayout
         android:id="@+id/students"
         android:layout_width="match_parent"
-        android:layout_height="100dp"
+        android:layout_height="150dp"
+        android:layout_marginTop="16dp"
         android:orientation="vertical"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.6"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/students_contacted"
-        app:layout_constraintVertical_bias="0.217">
+        app:layout_constraintTop_toBottomOf="@id/students_contacted">
 
     </LinearLayout>
 
@@ -47,12 +47,12 @@
         android:id="@+id/your_subjects"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="28dp"
-        android:layout_marginBottom="8dp"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="4dp"
         android:text="Your Subjects"
         android:textSize="24sp"
-        app:layout_constraintTop_toBottomOf="@id/students"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/students" />
 
     <LinearLayout
         android:id="@+id/subjects"
@@ -80,11 +80,11 @@
         android:id="@+id/your_history"
         android:layout_width="94dp"
         android:layout_height="32dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginBottom="4dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="24dp"
         android:text="Your History"
-        app:layout_constraintTop_toBottomOf="@id/subjects"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/subjects" />
 
     <TextView
         android:id="@+id/num_students_contacted"
