+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -19,7 +19,7 @@
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.example.hopportunities.ui.notifications.NotificationsFragment"
+        android:name="com.example.hopportunities.ui.notifications.StudentHomeFragment"
         android:label="@string/title_notifications"
         tools:layout="@layout/fragment_notifications" />
 </navigation>
